MonitoringFolder

+++ src/main/java/com/arthurmaciel/desafio/decode/FileDecoder.java
@@ -19,7 +19,7 @@ import com.arthurmaciel.desafio.exception.FolderDoesNotExistException;
 import com.arthurmaciel.desafio.exception.InvalidIDException;
 
 public class FileDecoder {
-	
+
 	private FileDAO fileDAO;
 	private Report report;
 	private ModelDAO modelDAO;
@@ -30,27 +30,45 @@ public class FileDecoder {
 	public FileDecoder() {
 		fileDAO = new FileDAO();
 	}
-	
+
 	public void decodeFile() {
 		for(String file : getFilesFromFolder()) {
-			List<String> lines = fileDAO.readFile(file);
-			report = new Report(getFileName(file));
-			modelDAO = new ModelDAO();
-			for(String line : lines) {
-				String[] split = line.split(SPLITLINE);
-				if(split.length == MAXSIZESPLIT) {
-					checkID(split);	
-				}									
+
+			if(!isFileDone(getFileName(file))) {
+
+				List<String> lines = fileDAO.readFile(file);
+				report = new Report(getFileName(file));
+				modelDAO = new ModelDAO();
+
+				for(String line : lines) {
+					String[] split = line.split(SPLITLINE);
+					if(split.length == MAXSIZESPLIT) {
+
+						try {
+							checkID(split);	
+						}catch(InvalidIDException e) {
+							System.out.println(e + "\n File:" + getFileName(file));
+						}
+
+					}					
+				}
+
+				decodeWorstSalesman();
+				mostExpensiveSale();
+
+				System.out.println(report);
+				fileDAO.writeFile(report);
 			}
-			
-			decodeWorstSalesman();
-			mostExpensiveSale();
-			
-			System.out.println(report);
-			fileDAO.writeFile(report);
 		}
 	}
-	
+
+	private boolean isFileDone(String fileName) {
+		if(fileDAO.getFilesDone().contains(fileName)) {
+			return true;
+		} 
+		return false;
+	}
+
 	private void checkID(String[] line) {
 
 		switch(line[0]) {
@@ -63,49 +81,49 @@ public class FileDecoder {
 		case "003":
 			decodeSale(line);
 			break;
-			default:
-				throw new InvalidIDException();
+		default:
+			throw new InvalidIDException();
 		}
 	}
-	
+
 	private void decodeSalesman(String[] salesman) {
 		modelDAO.addSalesman(new Salesman(salesman[1], salesman[2], salesman[3]));
 		report.addQtdSalesman();		
 	}
-	
+
 	private void decodeCostumer(String[] costumer) {
 		modelDAO.addCostumer(new Costumer(costumer[1], costumer[2], costumer[3]));
 		report.addQtdCostumers();
 	}
-	
+
 	private void decodeWorstSalesman() {
 		report.setWorstSalesman(modelDAO.getSalesmen().stream().reduce((salesman1, salesman2) -> 
-			salesSalesman(salesman1.getName()) < salesSalesman(salesman2.getName()) ? salesman1 : salesman2).get().getName());
+		salesSalesman(salesman1.getName()) < salesSalesman(salesman2.getName()) ? salesman1 : salesman2).get().getName());
 	}
-	
+
 	private void decodeSale(String[] sale) {
 		String salesmanName = sale[3];
 		Sale aux = new Sale(sale[1], findSalesman(salesmanName));
-		
+
 		String[] items = sale[2].replaceAll("\\[|\\]", "").split(",");
 
 		for(int i = 0; i < items.length ; i++) {
 			String[] itemDetail = items[i].split("-");
-			
+
 			aux.addItem(new Item(itemDetail[0], itemDetail[1], itemDetail[2]));
 		}
-		
+
 		modelDAO.addSale(aux);
 	}
-	
+
 	private Salesman findSalesman(String name) {
 		return modelDAO.getSalesmen().stream().filter(p -> p.getName().equals(name)).findFirst().get();
 	}
-	
+
 	private void mostExpensiveSale() {
 		report.setMostExpensiveSaleId(modelDAO.getSales().stream().reduce((sale1, sale2) -> sale1.getTotal() > sale2.getTotal() ? sale1 : sale2).get().getSaleId());
 	}
-	
+
 	private double salesSalesman(String salesmanName){
 		double total = 0;
 
@@ -116,25 +134,24 @@ public class FileDecoder {
 		}
 		return total;
 	}
-	
+
 	private List<String> getFilesFromFolder() {
 		try (Stream<Path> walk = Files.walk(Paths.get(FILEPATH))) {
 
-			List<String> result = walk.filter(Files::isRegularFile)
+			return walk.filter(Files::isRegularFile)
 					.map(x -> x.toString()).collect(Collectors.toList());
-			return result;
 
 		} catch (IOException e) {
 			throw new FolderDoesNotExistException();
 		}
 	}
-	
+
 	private String getFileName(String path) {
 		String[] aux = path.split("/");
 
-        String[] file = aux[aux.length-1].split("\\.");
+		String[] file = aux[aux.length-1].split("\\.");
 
-        return file[0];
+		return file[0];
 	}
 
 }
